Serve docs.forc.pub via src middleware
